pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mlsys-kv"
version = "0.1.0"
description = "CMU MLSys: joint KV sparsification and quantization for self-speculative decoding (course project)."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.2.0",
    "transformers>=4.40.0",
    "accelerate>=0.28.0",
    "sentencepiece>=0.2.0",
    "numpy>=1.26.0",
    "pandas>=2.0.0",
    "matplotlib>=3.8.0",
    "pyyaml>=6.0.0",
    "tqdm>=4.66.0",
    "modal>=0.64.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0", "ruff>=0.4.0"]

[project.scripts]
mlsys-kv = "mlsys_kv.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
